Tests/Docs after stricter explicit conversion.

+++ test/libsolidity/semanticTests/enums/enum_explicit_overflow.sol
@@ -9,12 +9,12 @@ contract test {
     }
 
     function getChoiceFromSigned(int256 x) public returns (uint256 d) {
-        choice = ActionChoices(x);
+        choice = ActionChoices(uint8(uint(x)));
         d = uint256(choice);
     }
 
     function getChoiceFromMax() public returns (uint256 d) {
-        choice = ActionChoices(type(uint).max);
+        choice = ActionChoices(uint8(type(uint).max));
         d = uint256(choice);
     }
